requirements.txt
requests
python-dotenv
nltk
pandas
numpy
matplotlib
seaborn
openpyxl
aiohttp